App is no more outside migrations and resources (#7190)
* App is no more ouside migrations and resources

* fix

* error handling

* nits and lints

* nit

* lint

* comment

* nits

* fix

+++ front/lib/resources/app_resource.ts
@@ -1,17 +1,17 @@
-import type { AppType, AppVisibility, Project, Result } from "@dust-tt/types";
+import type { AppType, AppVisibility, Result } from "@dust-tt/types";
 import { Err, Ok } from "@dust-tt/types";
 import type { Attributes, CreationAttributes, ModelStatic } from "sequelize";
 import { Op } from "sequelize";
 
 import type { Authenticator } from "@app/lib/auth";
-import type { DatasetResource } from "@app/lib/resources/dataset_resource";
+import { DatasetResource } from "@app/lib/resources/dataset_resource";
 import { ResourceWithVault } from "@app/lib/resources/resource_with_vault";
 import { RunResource } from "@app/lib/resources/run_resource";
 import { frontSequelize } from "@app/lib/resources/storage";
 import { App, Clone } from "@app/lib/resources/storage/models/apps";
 import type { ReadonlyAttributesType } from "@app/lib/resources/storage/types";
-import { generateLegacyModelSId } from "@app/lib/resources/string_ids";
-import { VaultResource } from "@app/lib/resources/vault_resource";
+import type { ResourceFindOptions } from "@app/lib/resources/types";
+import type { VaultResource } from "@app/lib/resources/vault_resource";
 
 // Attributes are marked as read-only to reflect the stateless nature of our Resource.
 // This design will be moved up to BaseResource once we transition away from Sequelize.
@@ -21,16 +21,12 @@ export interface AppResource extends ReadonlyAttributesType<App> {}
 export class AppResource extends ResourceWithVault<App> {
   static model: ModelStatic<App> = App;
 
-  private datasets: DatasetResource[];
-
   constructor(
     model: ModelStatic<App>,
     blob: Attributes<App>,
-    vault: VaultResource,
-    datasets: DatasetResource[]
+    vault: VaultResource
   ) {
-    super(AppResource.model, blob, vault);
-    this.datasets = datasets;
+    super(App, blob, vault);
   }
 
   static async makeNew(
@@ -42,47 +38,130 @@ export class AppResource extends ResourceWithVault<App> {
       vaultId: vault.id,
     });
 
-    return new this(AppResource.model, app.get(), vault, []);
+    return new this(App, app.get(), vault);
   }
 
-  // Cloning.
+  // Fetching.
 
-  async clone({
-    targetAuth,
-    name,
-    description,
-    visibility,
-    coreProject,
-  }: {
-    targetAuth: Authenticator;
-    name: string;
-    description: string | null;
-    visibility: AppVisibility;
-    coreProject: Project;
-  }) {
-    // TODO: still WIP
-    const targetGlobalVault =
-      await VaultResource.fetchWorkspaceGlobalVault(targetAuth);
+  private static async baseFetch(
+    auth: Authenticator,
+    options?: ResourceFindOptions<App>
+  ) {
+    const apps = await this.baseFetchWithAuthorization(auth, {
+      ...options,
+    });
 
-    const cloned = AppResource.makeNew(
+    return apps.filter(
+      (app) => auth.isAdmin() || auth.hasPermission([app.vault.acl()], "read")
+    );
+  }
+
+  // fetchByIds filters out private apps if the auth is not a user on the workspace. This will be
+  // removed soon as we remove public apps.
+  static async fetchByIds(
+    auth: Authenticator,
+    ids: string[]
+  ): Promise<AppResource[]> {
+    return this.baseFetch(auth, {
+      where: {
+        sId: ids,
+        visibility: auth.isUser() ? ["public", "private"] : ["public"],
+      },
+    });
+  }
+
+  static async fetchById(
+    auth: Authenticator,
+    id: string
+  ): Promise<AppResource | null> {
+    const [app] = await this.fetchByIds(auth, [id]);
+
+    return app ?? null;
+  }
+
+  // listByWorkspace filters out private apps if the auth is not a user on the workspace. This will
+  // be removed soon as we remove public apps.
+  static async listByWorkspace(auth: Authenticator) {
+    return this.baseFetch(auth, {
+      where: {
+        workspaceId: auth.getNonNullableWorkspace().id,
+        visibility: auth.isUser() ? ["public", "private"] : ["public"],
+      },
+    });
+  }
+
+  // Mutation.
+
+  async updateState(
+    auth: Authenticator,
+    {
+      savedSpecification,
+      savedConfig,
+      savedRun,
+    }: {
+      savedSpecification: string;
+      savedConfig: string;
+      savedRun?: string;
+    }
+  ) {
+    await this.model.update(
       {
-        sId: generateLegacyModelSId(),
+        savedSpecification,
+        savedConfig,
+        savedRun,
+      },
+      {
+        where: {
+          id: this.id,
+          workspaceId: auth.getNonNullableWorkspace().id,
+        },
+      }
+    );
+  }
+
+  async updateSettings(
+    auth: Authenticator,
+    {
+      name,
+      description,
+      visibility,
+    }: {
+      name: string;
+      description: string | null;
+      visibility: AppVisibility;
+    }
+  ) {
+    await this.model.update(
+      {
         name,
         description,
         visibility,
-        dustAPIProjectId: coreProject.project_id.toString(),
-        savedSpecification: this.savedSpecification,
-        workspaceId: targetAuth.getNonNullableWorkspace().id,
       },
-      targetGlobalVault
+      {
+        where: {
+          id: this.id,
+          workspaceId: auth.getNonNullableWorkspace().id,
+        },
+      }
     );
+  }
 
-    return cloned;
+  async markAsDeleted(auth: Authenticator) {
+    await this.model.update(
+      {
+        visibility: "deleted",
+      },
+      {
+        where: {
+          id: this.id,
+          workspaceId: auth.getNonNullableWorkspace().id,
+        },
+      }
+    );
   }
 
   // Deletion.
 
-  // TODO: not yet used
   async delete(auth: Authenticator): Promise<Result<undefined, Error>> {
     try {
       await frontSequelize.transaction(async (t) => {
@@ -93,7 +172,11 @@ export class AppResource extends ResourceWithVault<App> {
           },
           transaction: t,
         });
-        await Promise.all(this.datasets.map((d) => d.delete(auth, t)));
+        const res = await DatasetResource.deleteForApp(auth, this, t);
+        if (res.isErr()) {
+          // Interrupt the transaction if there was an error deleting datasets.
+          throw res.error;
+        }
         await this.model.destroy({
           where: {
             workspaceId: auth.getNonNullableWorkspace().id,
